src/edu/grinnell/csc2907/wolterzo/utils/StringUtilsTest.java
package edu.grinnell.csc2907.wolterzo.utils;

import static org.junit.Assert.*;

import org.junit.Test;

public class StringUtilsTest
{

  @Test
  public void testSplitAt()
  {
    assertArrayEquals(new String[] { "a", "b", "c" },
                      StringUtils.splitAt("a:b:c", ':'));
    assertArrayEquals(new String[] { "a", "b", "c" },
                      StringUtils.splitAt("a b c", ' '));
    assertArrayEquals(new String[] { "a:b:c" },
                      StringUtils.splitAt("a:b:c", ' '));
    assertArrayEquals("one field", new String[] { "a" },
                      StringUtils.splitAt("a", ':'));
    assertArrayEquals("empty inner field", new String[] { "a", "", "c" },
                      StringUtils.splitAt("a::c", ':'));
    assertArrayEquals("leading empty field", new String[] { "", "a" },
                      StringUtils.splitAt(":a", ':'));
    assertArrayEquals("trailing empty field", new String[] { "a", "" },
                      StringUtils.splitAt("a:", ':'));

  }
  
  @Test
  public void testSplitCSV()
  {
    assertArrayEquals(new String[] { "a", "b", "c" },
                      StringUtils.splitCSV("a,b,c"));
    assertArrayEquals(new String[] { "a,b", "c" },
                      StringUtils.splitCSV("\"a,b\",c"));
    assertArrayEquals(new String[] { "a", "b,b\"", "c" },
                      StringUtils.splitCSV("a,\"b,b\"\"\",c"));
  }

}
